requirements.txt
astroid==2.3.2
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pdfminer.six==20191020
pycryptodome==3.9.0
pylint==2.4.3
requests==2.22.0
schedule==0.6.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.4
typed-ast==1.4.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
